+++ CHANGES.md
@@ -8,6 +8,9 @@ Version 0.14.5
 
 To be released.
 
+ -  Fixed a bug of `Object.toJsonLd()` method where it had not fall back to
+    the proper compact form when the heuristic compact form was not available.
+
 
 Version 0.14.4
 --------------